fix: route Spark UTF8 binary formatter through JVM-compatible decoder
Addresses review feedback on the CAST(binary AS string) UB fix:

- BinaryOutputStyle::Utf8 previously rendered via
  String::from_utf8(..).unwrap(), panicking the executor on non-UTF-8
  input. With spark.sql.binaryOutputStyle=UTF8 (Spark 4.0+), a
  ToPrettyString/show() over a binary column with invalid bytes hit
  this. Decode lossily instead so invalid bytes become U+FFFD, matching
  Spark's new String(bytes, UTF_8).
- Build cast_binary_to_string with a GenericStringBuilder and append the
  decoder's borrowed Cow, so the common valid-UTF-8 path copies once
  instead of allocating a throwaway String and copying twice.
- Move decode_utf8_spark_lossy into datafusion-comet-common next to
  bytes_to_i128 so the shuffle crate no longer reaches into spark-expr.
- Document why the char::from_u32 unwraps in the decoder are infallible.
