+++ native/spark-expr/src/conversion_funcs/cast.rs
@@ -36,13 +36,11 @@ use crate::conversion_funcs::temporal::{
     is_df_cast_from_timestamp_spark_compatible,
 };
 use crate::conversion_funcs::utils::spark_cast_postprocess;
-use crate::utils::{
-    array_with_timezone, cast_timestamp_to_ntz, decode_utf8_spark_lossy, timestamp_ntz_to_timestamp,
-};
+use crate::utils::{array_with_timezone, cast_timestamp_to_ntz, timestamp_ntz_to_timestamp};
 use crate::EvalMode::Legacy;
 use crate::{cast_whole_num_to_binary, BinaryOutputStyle};
 use crate::{EvalMode, SparkError};
-use arrow::array::builder::StringBuilder;
+use arrow::array::builder::{GenericStringBuilder, StringBuilder};
 use arrow::array::{
     new_null_array, BinaryBuilder, DictionaryArray, GenericByteArray, ListArray, MapArray,
     StringArray, StructArray,
@@ -52,8 +50,8 @@ use arrow::datatypes::{Field, Fields, GenericBinaryType};
 use arrow::error::ArrowError;
 use arrow::{
     array::{
-        cast::AsArray, types::Int32Type, Array, ArrayRef, GenericStringArray, Int16Array,
-        Int32Array, Int64Array, Int8Array, OffsetSizeTrait, PrimitiveArray,
+        cast::AsArray, types::Int32Type, Array, ArrayRef, Int16Array, Int32Array, Int64Array,
+        Int8Array, OffsetSizeTrait, PrimitiveArray,
     },
     compute::{cast_with_options, take, CastOptions},
     record_batch::RecordBatch,
@@ -64,6 +62,8 @@ use base64::Engine;
 use datafusion::common::{internal_err, DataFusionError, Result as DataFusionResult, ScalarValue};
 use datafusion::physical_expr::PhysicalExpr;
 use datafusion::physical_plan::ColumnarValue;
+use datafusion_comet_common::decode_utf8_spark_lossy;
+use std::borrow::Cow;
 use std::{
     fmt::{Debug, Display, Formatter},
     hash::Hash,
@@ -803,21 +803,22 @@ fn cast_binary_to_string<O: OffsetSizeTrait>(
         .downcast_ref::<GenericByteArray<GenericBinaryType<O>>>()
         .unwrap();
 
-    fn binary_formatter(value: &[u8], spark_cast_options: &SparkCastOptions) -> String {
-        match spark_cast_options.binary_output_style {
-            Some(s) => spark_binary_formatter(value, s),
-            None => cast_binary_formatter(value),
+    // Build with a GenericStringBuilder and append &str straight from the decoder's Cow so the
+    // common valid-UTF-8 cast path copies the bytes once (into the Arrow value buffer) instead of
+    // twice (an owned String, then a copy into the buffer).
+    let mut builder = GenericStringBuilder::<O>::new();
+    for value in input.iter() {
+        match value {
+            Some(value) => match spark_cast_options.binary_output_style {
+                // ToPrettyString styles (Spark 4.0+) build owned strings, which is unavoidable.
+                Some(s) => builder.append_value(spark_binary_formatter(value, s)),
+                // Default CAST(binary AS string): borrows in the valid path, appends once.
+                None => builder.append_value(cast_binary_formatter(value)),
+            },
+            None => builder.append_null(),
         }
     }
-
-    let output_array = input
-        .iter()
-        .map(|value| match value {
-            Some(value) => Ok(Some(binary_formatter(value, spark_cast_options))),
-            _ => Ok(None),
-        })
-        .collect::<Result<GenericStringArray<O>, ArrowError>>()?;
-    Ok(Arc::new(output_array))
+    Ok(Arc::new(builder.finish()))
 }
 
 /// This function mimics the [BinaryFormatter]: https://github.com/apache/spark/blob/v4.0.0/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/ToStringBase.scala#L449-L468
@@ -826,7 +827,10 @@ fn cast_binary_to_string<O: OffsetSizeTrait>(
 /// Before Spark 4.0.0, the default is SPACE_DELIMITED_UPPERCASE_HEX
 fn spark_binary_formatter(value: &[u8], binary_output_style: BinaryOutputStyle) -> String {
     match binary_output_style {
-        BinaryOutputStyle::Utf8 => String::from_utf8(value.to_vec()).unwrap(),
+        // Spark's UTF8 BinaryFormatter renders via `UTF8String.fromBytes(bytes).toString`, i.e.
+        // `new String(bytes, UTF_8)`. Route through the shared JVM-compatible decoder so invalid
+        // bytes become U+FFFD (matching Spark) instead of panicking on non-UTF-8 input (#4488).
+        BinaryOutputStyle::Utf8 => decode_utf8_spark_lossy(value).into_owned(),
         BinaryOutputStyle::Basic => {
             format!(
                 "{:?}",
@@ -855,7 +859,7 @@ fn spark_binary_formatter(value: &[u8], binary_output_style: BinaryOutputStyle) 
     }
 }
 
-fn cast_binary_formatter(value: &[u8]) -> String {
+fn cast_binary_formatter(value: &[u8]) -> Cow<'_, str> {
     // CAST(binary AS string) reinterprets the bytes as UTF-8, like Spark's UTF8String.fromBytes.
     // Spark keeps the raw bytes, but Arrow's Utf8 type requires valid UTF-8, and building a String
     // from non-UTF-8 bytes is undefined behaviour (#4488). Decode JVM-compatibly-lossily instead:
@@ -864,8 +868,9 @@ fn cast_binary_formatter(value: &[u8]) -> String {
     // result is memory-safe valid UTF-8, never feeds invalid bytes into downstream native string
     // kernels, and matches Spark's rendered output byte-for-byte. It diverges from Spark only under
     // byte-level round-trips such as CAST(CAST(x AS string) AS binary), where Spark still has the
-    // original bytes and Comet has the U+FFFD replacements.
-    decode_utf8_spark_lossy(value).into_owned()
+    // original bytes and Comet has the U+FFFD replacements. Returning `Cow` lets the valid path
+    // borrow so the caller appends without an intermediate allocation.
+    decode_utf8_spark_lossy(value)
 }
 
 #[cfg(test)]
@@ -890,6 +895,31 @@ mod tests {
         ]);
         // binary_output_style defaults to None, i.e. the plain (non-ToPrettyString) cast path.
         let cast_options = SparkCastOptions::new(EvalMode::Legacy, "UTC", false);
+
+        let result = cast_binary_to_string::<i32>(&input, &cast_options).unwrap();
+
+        let strings = result.as_string::<i32>();
+        assert_eq!(strings.len(), 4);
+        assert_eq!(strings.value(0), "\u{FFFD}\u{FFFD}");
+        assert!(strings.is_null(1));
+        assert_eq!(strings.value(2), "abc");
+        assert_eq!(strings.value(3), "\u{FFFD}");
+    }
+
+    #[test]
+    fn test_cast_binary_to_string_utf8_output_style_replaces_invalid_utf8() {
+        // Spark's `binaryOutputStyle=UTF8` (Spark 4.0+) ToPrettyString formatter renders binary via
+        // `new String(bytes, UTF_8)`. Previously this arm called `String::from_utf8(..).unwrap()`,
+        // which panicked the executor on non-UTF-8 input. It must now decode JVM-compatibly-lossily,
+        // matching the default cast path's replacement behavior.
+        let input = BinaryArray::from_opt_vec(vec![
+            Some(&[0xFFu8, 0xFE][..]),
+            None,
+            Some("abc".as_bytes()),
+            Some(&[0xEDu8, 0xA0, 0x80][..]),
+        ]);
+        let mut cast_options = SparkCastOptions::new(EvalMode::Legacy, "UTC", false);
+        cast_options.binary_output_style = Some(BinaryOutputStyle::Utf8);
 
         let result = cast_binary_to_string::<i32>(&input, &cast_options).unwrap();
